Fixed warning in ExtSenderProcessor, and reordered the DDxmlParser ifs
GetChannelNumber() and GetModuleNumber() both will never be bigger than an int so the cast is ok. Really all this does is suppress a warning when compiling

alphabetized the if statements in the DetectorDriverXmlParser: makes it easier to read

+++ Analysis/Utkscan/processors/source/ExtTSSenderProcessor.cpp
@@ -47,7 +47,7 @@ bool ExtTSSenderProcessor::Process(RawEvent &event)
 
     static const std::vector<ChanEvent *> &chEvents = event.GetEventList();
     for( auto chEvent : chEvents ) {
-        if( chEvent->GetChannelNumber() == chanNum_ && chEvent->GetModuleNumber() == slotNum_ ){
+        if( (int)chEvent->GetChannelNumber() == chanNum_ && (int)chEvent->GetModuleNumber() == slotNum_ ){
             unsigned long long ts = chEvent->GetExternalTimeStamp();
             // printf("ts %llu \n", ts);
             SetTS(ts);
